moved actions to corresponding file
