+++ app/react/Metadata/actions/MetadataExtractionActions.ts
@@ -1,0 +1,19 @@
+import { actions } from 'app/BasicReducer';
+import { actions as formActions } from 'react-redux-form';
+
+//should only occur when form is loaded
+const setSelectedProperties = async () => {};
+
+const selectionHandler = (selection: {}, fieldName: string, fieldId?: string) => {
+  const data = {
+    ...(fieldId && { _id: fieldId }),
+    label: fieldName,
+    timestamp: Date(),
+    selection,
+  };
+  return actions.updateIn('documentViewer.metadataExtraction', ['selections'], data);
+};
+
+const formFieldUpdater = (value: string, model: string) => formActions.change(model, value);
+
+export { selectionHandler, formFieldUpdater };
